feat: Implement CLI commands using Typer and add version/settings functionality
