+++ src/pymmcore_gui/_app.py
@@ -143,11 +143,11 @@ def _close_splash_screen() -> None:  # pragma: no cover
 
 
 def _decide_configuration(
-    config: str | None = None, parent: QWidget | None = None
+    config: Path | str | None = None, parent: QWidget | None = None
 ) -> str | None:
     settings = Settings.instance()
     if config:
-        return config
+        return str(config)
 
     if last_config := settings.last_config:
         if auto_load := settings.auto_load_last_config:
